rebase(): rebase à partir de master, et refacto des tests
