--- src/main/java/com/example/springbootformation/domain/Film.java
@@ -1,7 +1,0 @@
-package com.example.springbootformation.domain;
-
-import java.util.Date;
-import java.util.List;
-import java.util.UUID;
-
-public record Film(UUID id, String nom, String synopsis, List<String> genres, Date dateDeSortie) {}
