--- src/main/java/com/example/springbootformation/domain/FilmVu.java
@@ -1,5 +1,0 @@
-package com.example.springbootformation.domain;
-
-import java.util.UUID;
-
-public record FilmVu(UUID utilisateurId, UUID filmId, Float note, String commentaire) {}
