--- src/main/java/com/example/springbootformation/domain/Utilisateur.java
@@ -1,5 +1,0 @@
-package com.example.springbootformation.domain;
-
-import java.util.UUID;
-
-public record Utilisateur(UUID id, String nom, String prenom) {}
