--- src/main/java/com/example/springbootformation/domain/repositories/UtilisateurRepository.java
@@ -1,7 +1,0 @@
-package com.example.springbootformation.domain.repositories;
-
-import com.example.springbootformation.domain.Utilisateur;
-
-public interface UtilisateurRepository {
-  Utilisateur recupererUtilisateurConnecte();
-}
