+++ src/main/java/com/octo/ajava/infra/controllers/ApiController.java
@@ -1,15 +1,15 @@
-package com.example.springbootformation.infra.controller;
+package com.octo.ajava.infra.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/accueil")
-public class AccueilController {
+@RequestMapping("/api/")
+public class ApiController {
 
   @GetMapping
-  public String accueil() {
-    return "Bienvenue à la formation AJAVA.";
+  public String presentationDeLApi() {
+    return "Présentation de l'API Films";
   }
 }
