--- src/test/java/com/example/springbootformation/SpringbootformationApplicationTests.java
@@ -1,11 +1,0 @@
-package com.example.springbootformation;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringbootformationApplicationTests {
-
-  @Test
-  void contextLoads() {}
-}
