+++ src/test/java/com/octo/ajava/ApiIntegrationTest.java
@@ -1,16 +1,18 @@
-package com.example.springbootformation;
+package com.octo.ajava;
 
-import static com.example.springbootformation.SpringbootformationApplication.FORMATEUR_DATE;
+import static com.octo.ajava.AjavaApplication.FORMATEUR_DATE;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.example.springbootformation.infra.security.WebSecurityConfiguration;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
+import com.octo.ajava.infra.security.WebSecurityConfiguration;
+import java.nio.charset.StandardCharsets;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 
@@ -34,5 +36,10 @@ public abstract class ApiIntegrationTest {
 
   protected void laRouteEstProtegee(MockHttpServletRequestBuilder route) throws Exception {
     mockMvc.perform(route).andExpect(status().isUnauthorized());
+  }
+
+  protected <T> T convertirJsonEnObjet(MockHttpServletResponse response, Class<T> clazz)
+      throws Exception {
+    return objectMapper.readValue(response.getContentAsString(StandardCharsets.UTF_8), clazz);
   }
 }
